# ENH Improves efficiency of ColumnTransformer for string keys