--- a/sklearn/utils/__init__.py
+++ b/sklearn/utils/__init__.py
@@ -442,9 +442,9 @@ def _get_column_indices(X, key):
             column_indices = []
             for col in columns:
                 col_idx = all_columns.get_loc(col)
-                if not isinstance(col_idx, int):
-                    raise ValueError("Selected columns, {}, are not unique "
-                                     "in dataframe".format(columns))
+                if not isinstance(col_idx, numbers.Integer):
+                    raise ValueError(f"Selected columns, {columns}, are not "
+                                     "unique in dataframe")
                 column_indices.append(col_idx)
 
         except KeyError as e: