<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/scikit-learn/scikit-learn/blob/master/CONTRIBUTING.md#pull-request-checklist
-->

#### Reference Issues/PRs
<!--
Example: Fixes #1234. See also #3456.
Please use keywords (e.g., Fixes) to create link to the issues or pull requests
you resolved, so that they will automatically be closed when your pull request
is merged. See https://github.com/blog/1506-closing-issues-via-pull-requests
-->
Fixes https://github.com/scikit-learn/scikit-learn/issues/16327


#### What does this implement/fix? Explain your changes.
Uses the pandas index to find column indices

#### Any other comments?
Is a benchmark for this PR

```py
from sklearn.preprocessing import FunctionTransformer
from sklearn.compose import make_column_transformer
import numpy as np
import pandas as pd

n_features = 10_000
selected_features = 7_000
columns = [f'col_{i}' for i in range(n_features)]
df = pd.DataFrame(np.ones((1, n_features)), columns=columns)

ct = make_column_transformer(*[(FunctionTransformer(), [col])
                               for col in columns[:selected_features]])
```

### This PR

```
%%timeit
ct.fit_transform(df)
# 8.13 s ± 76 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### master

```
%%timeit
ct.fit_transform(df)
# 16.4 s ± 62 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```